plugins/amdgpu: Implement parallel restore
This patch implements the entire logic to enable the offloading of
buffer object content restoration. It has two parts: the first replaces
the restoration of buffer objects in the target process by sending a
parallel restore command to the main CRIU process; the second implements
the `RESTORE_ASYNCHRONOUS` hook in the amdgpu plugin to enable buffer
object content restoration in the main CRIU process.

Signed-off-by: Yanning Yang <yangyanning@sjtu.edu.cn>