cr-restore: Make `restore_wait_inprogress_tasks` external
Currently, `restore_wait_inprogress_tasks` is a static function and can
only be called within `cr-restore.c`. However, to implement parallel
restore, amdgpu plugin also needs to check the tasks' state to decide whether
to stop the parallel restore server. Therefore, this patch moves the
declaration of `restore_wait_inprogress_tasks` to `restore.h` so that it can
be called by the plugin.

Signed-off-by: Yanning Yang <yangyanning@sjtu.edu.cn>